move tpch tests into same file
