--- src/Database.Test/TPCH/Query03.cs
@@ -1,16 +1,0 @@
-using Database.Core.Execution;
-using FluentAssertions;
-
-namespace Database.Test.TPCH;
-
-public partial class TPCHTests
-{
-    [Test]
-    public void Q03()
-    {
-        var query = ReadQuery("query_03.sql");
-
-        var result = Query(query).AsRowList();
-        result.Should().HaveCountGreaterOrEqualTo(1);
-    }
-}
