--- src/Database.Test/TPCH/Query18.cs
@@ -1,19 +1,0 @@
-using Database.Core;
-using Database.Core.BufferPool;
-using Database.Core.Catalog;
-using Database.Core.Execution;
-using Database.Core.Planner;
-using FluentAssertions;
-
-namespace Database.Test.TPCH;
-
-public partial class TPCHTests
-{
-    [Test]
-    public void Q18()
-    {
-        var query = ReadQuery("query_18.sql");
-        var result = Query(query).AsRowList();
-        result.Should().HaveCountGreaterOrEqualTo(1);
-    }
-}
